+++ tablo_legacy_m3u/__init__.py
@@ -20,6 +20,7 @@ def create_app(
     config: "Config",
     tablo_client: "TabloClient",
     server_info: "ServerInfo",
+    enable_epg: bool,
 ) -> Flask:
     """Create and configure the Flask application."""
     app = Flask(__name__)
@@ -28,6 +29,7 @@ def create_app(
         APP_CONFIG=config,
         TABLO_CLIENT=tablo_client,
         TABLO_SERVER_INFO=server_info,
+        ENABLE_EPG=enable_epg,
     )
 
     register_routes(app)