+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 - Add `device.xml` endpoint and extract discover module <https://github.com/gtronset/tablo-legacy-m3u/pull/6>
 - Improve README <https://github.com/gtronset/tablo-legacy-m3u/pull/7>
 - Add status page at root and djlint linting <https://github.com/gtronset/tablo-legacy-m3u/pull/8>
+- Add XMLTV EPG endpoint <https://github.com/gtronset/tablo-legacy-m3u/pull/9>
 
 ## [0.1.0] - 2026-03-26